requirements.txt
streamlit==1.26.0
requests
python-dotenv
google-generativeai

